==> prediction/2-Preproc-Adri.ipynb <==
import matplotlib.pyplot as plt
from matplotlib.colors import LogNorm
from matplotlib.mlab import bivariate_normal
import pandas as pd
from dateutil.parser import parse
import numpy as np
import warnings
import datetime as dt
import seaborn as sns
warnings.filterwarnings('ignore')
import plotly

# Set de Entrenamiento de Postulaciones  

avisos_detalles = pd.read_csv('../data/adri_data/avisos_detalles_2.csv')
avisos_detalles.head()

avisos_detalles.drop(['titulo','descripcion', 'contenido',                     
'idzona_int',
'iddenominacion_empresa_int',
'idciudad_int',
'idmapacalle_int'], axis=1, inplace=True)
avisos_detalles.head()

postulaciones = pd.read_csv('../data/adri_data/postulaciones.csv')
postulaciones.head()

postulaciones = postulaciones.drop('year', axis=1)

postulaciones_sin_descripcion = pd.merge(avisos_detalles, postulaciones, how='inner', on='idaviso')
postulaciones_sin_descripcion.head()

postulantes = pd.read_csv('../data/adri_data/postulantes3.csv')
postulantes.head()

postulantes.drop(['idnombre_area_int_pref','idnivel_laboral_int_pref','idtipo_de_trabajo_int_pref', 'idpostulante_int'], inplace=True, axis=1)

postulaciones_sin_descripcion = pd.merge(postulaciones_sin_descripcion, postulantes, how='inner', on='idpostulante')
postulaciones_sin_descripcion.count()

postulaciones_sin_descripcion.to_csv('../data/adri_data/entrenamiento_postulaciones.csv', sep=',', encoding='utf-8', index=False)
postulaciones_sin_descripcion.head()

# Set de Entrenamiento de Vistas

vistas = pd.read_csv('../data/adri_data/vistas.csv')
vistas.count()

# vistas.drop_duplicates(['idaviso', 'idpostulante'], inplace=True)
# vistas.count()

vistas.head()

vistas_con_detalle = pd.merge(vistas, postulantes, how='inner', on='idpostulante')
vistas_con_detalle.count()

vistas_con_detalle = pd.merge(vistas_con_detalle, avisos_detalles, how='inner', on='idaviso')
vistas_con_detalle.count()

vistas_con_detalle.head()

# vistas_con_detalle.drop('year', inplace=True, axis=1)

vistas_con_detalle.to_csv('../data/adri_data/entrenamiento_vistas2.csv', sep=',', encoding='utf-8', index=False)
vistas_con_detalle.head()
