==> prediction/3-Adri-merge.ipynb <==
import matplotlib.pyplot as plt
from matplotlib.colors import LogNorm
from matplotlib.mlab import bivariate_normal
import pandas as pd
from dateutil.parser import parse
import numpy as np
import warnings
import datetime as dt
import seaborn as sns
warnings.filterwarnings('ignore')

from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LogisticRegression
from sklearn import metrics

vistas = pd.read_csv('../data/adri_data/entrenamiento_vistas2.csv');
vistas.head()

postulaciones = pd.read_csv('../data/adri_data/entrenamiento_postulaciones2.csv');
postulaciones.head()

postulaciones.columns

vistas = vistas[postulaciones.columns]

postulaciones = postulaciones[postulaciones.columns]

vistas['se_postulo'] = 0
postulaciones['se_postulo'] = 1

vistas.shape

postulaciones.shape

vistas_y_postulaciones = pd.concat([postulaciones, vistas], axis=0)
vistas_y_postulaciones.shape

vistas_y_postulaciones.head()

# Agrego la tercer clase

top100 = pd.read_csv("../test_final_100k.csv", encoding='utf-8')
top100.head()

top100_full = pd.merge(top100, vistas_y_postulaciones, how='left', on=['idaviso', 'idpostulante'])
top100_full.head()

excluded = top100_full[top100_full['esta_online'].isnull()]
excluded.shape

excluded = excluded[['id', 'idaviso', 'idpostulante']]
excluded.head()

avisos_detalles = pd.read_csv('../data/adri_data/avisos_detalles.csv', encoding='utf-8')
avisos_detalles.head()

excluded = pd.merge(excluded, avisos_detalles, how='inner', on='idaviso')
excluded.shape

postulantes = pd.read_csv('../data/adri_data/postulantes2.csv', encoding='utf-8')
postulantes.shape

excluded = pd.merge(excluded, postulantes, how='inner', on='idpostulante')
excluded.shape

excluded['se_postulo'] = -1

excluded['month'] = 0
excluded['day'] = 0
excluded['dayofweek'] = 0
excluded['hour'] = 0

excluded['cantidad_postulaciones_x'] = 0
excluded['cantidad_vistas_x'] = 0
excluded['cantidad_postulaciones_y'] = 0
excluded['cantidad_vistas_y'] = 0
excluded['cantidad_interacciones'] = 0

excluded = excluded[vistas_y_postulaciones.columns]
excluded.head()

excluded = excluded[vistas_y_postulaciones.columns]

# timestamps = pd.DataFrame(vistas_y_postulaciones.groupby('idaviso')['timestamp'].mean()).reset_index()
# timestamps.head()

# excluded = pd.merge(excluded, timestamps, on='idaviso', how='left')
# excluded.head()

# excluded = excluded[vistas_y_postulaciones.columns]
# excluded.head()

# timestampmean = vistas_y_postulaciones['timestamp'].mean()
# timestampmean

# excluded.loc[excluded['timestamp'].isnull(), 'timestamp'] = timestampmean
# excluded.head()

# excluded[excluded['timestamp'].isnull()].head()

# excluded.loc[excluded['timestamp'].isnull(), 'timestamp'] = timestampmean
# excluded.head()

# Mergeo con la tercer clase

vistas_y_postulaciones = pd.concat([vistas_y_postulaciones, excluded], axis=0)
vistas_y_postulaciones.head()

# vistas_y_postulaciones.shape

# postulantes_ids = vistas_y_postulaciones[["idpostulante", "idpostulante2"]].drop_duplicates()
# postulantes_ids.to_csv('../data/adri_data/entrenamiento_id_postulantes.csv', encoding='utf-8', index=False)

# vistas_y_postulaciones.drop('idpostulante', axis=1, inplace=True)
# vistas_y_postulaciones.rename({'idpostulante_int': 'idpostulante'}, inplace=True, axis=1)
# vistas_y_postulaciones.head()

# vistas_y_postulaciones.to_csv('../data/adri_data/entrenamiento_postulaciones_vistas2.csv', encoding='utf-8', index=False)

# Luego de mergear

# vistas_y_postulaciones = pd.read_csv('../data/adri_data/entrenamiento_postulaciones_vistas2.csv', encoding='utf-8')

vistas_y_postulaciones.shape

vistas_y_postulaciones.head()

# dup = vistas_y_postulaciones.duplicated(['idaviso', 'idpostulante'])
# vistas_y_postulaciones[dup].shape

# vistas_y_postulaciones.drop_duplicates(['idaviso', 'idpostulante'], inplace=True)
# vistas_y_postulaciones.shape

vistas_y_postulaciones = vistas_y_postulaciones.sample(frac=1).reset_index(drop=True)

vistas_y_postulaciones.shape

vistas_y_postulaciones.info(memory_usage='deep')

# vistas_y_postulaciones.drop(['idnombre_area_int_pref','idnivel_laboral_int_pref','idtipo_de_trabajo_int_pref'], inplace=True, axis=1)

vistas_y_postulaciones.head()

vistas_y_postulaciones.to_csv('../data/adri_data/entrenamiento_postulaciones_vistas19_full.csv', encoding='utf-8', index=False)
