==> prediction/4-Adri KNN.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt
from sklearn import linear_model
from sklearn import model_selection
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
import matplotlib.pyplot as plt
import seaborn as sb
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LogisticRegression
from sklearn import metrics

from sklearn.cross_validation import train_test_split
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import accuracy_score
from sklearn.decomposition import PCA
from sklearn import preprocessing

import warnings
from math import log
import datetime as dt
import seaborn as sns
warnings.filterwarnings('ignore')

df = pd.read_csv('../data/adri_data/entrenamiento_postulaciones_vistas19_4_pp.csv', encoding='utf-8')
df.shape

df['se_postulo'] = df['se_postulo'].apply(lambda x: 0 if x == 0 else 1)

df['sepostulo'] = 0

df = pd.concat([df, df[df['se_postulo'] == 0]], axis=0)

# df['cantidad_interacciones_y'] = df['cantidad_interacciones_y'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_interacciones_x'] = df['cantidad_interacciones_x'].apply(lambda x: 0 if x < 1 else log(x))
df['cantidad_postulaciones_y'] = df['cantidad_postulaciones_y'].apply(lambda x: 0 if x < 1 else log(x))
df['cantidad_postulaciones_x'] = df['cantidad_postulaciones_x'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_vistas_y'] = df['cantidad_vistas_y'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_vistas_x'] = df['cantidad_vistas_x'].apply(lambda x: 0 if x < 1 else log(x))

df['cantidad_vistas_y'] = df['cantidad_vistas_y'].apply(lambda x: 0 if x < 1 else log(x))
df['cantidad_vistas_x'] = df['cantidad_vistas_x'].apply(lambda x: 0 if x < 1 else log(x))

# df['cantidad_interacciones_y'] = df['cantidad_interacciones_y'].apply(lambda x: 0 if x <= 0 else 1)
# df['cantidad_interacciones_x'] = df['cantidad_interacciones_x'].apply(lambda x: 0 if x <= 0 else 1)
# df['cantidad_postulaciones_y'] = df['cantidad_postulaciones_y'].apply(lambda x: 0 if x <= 0 else 1)
# df['cantidad_postulaciones_x'] = df['cantidad_postulaciones_x'].apply(lambda x: 0 if x <= 0 else 1)
# df['cantidad_vistas_y'] = df['cantidad_vistas_y'].apply(lambda x: x if x <= 0 else 1)
# df['cantidad_vistas_x'] = df['cantidad_vistas_x'].apply(lambda x: x if x <= 0 else 1)

df['post_tipo_de_trabajo'] = df['post_tipo_de_trabajo'].apply(lambda x: 0 if x <= 0 else log(x))
df['post_nivel_laboral'] = df['post_nivel_laboral'].apply(lambda x: 0 if x <= 0 else log(x))
df['post_nombre_area'] = df['post_nombre_area'].apply(lambda x: 0 if x <= 0 else log(x))

# df.drop(['idzona', 'idciudad','idmapacalle'], axis=1, inplace=True)
df.head()

df['valor_educacion'] = df['valor_educacion'].astype('int')
df['educacion_int'] = df['educacion_int'].astype('int')
df['esta_online'] = df['esta_online'].apply(lambda x: 1 if x == 1 else 0)
# df['se_postulo'] = df['se_postulo'].apply(lambda x: 1 if x == 1 else -1)

cols = df.head(0)
cols.drop(['se_postulo', 'idpostulante', 'sepostulo'], inplace=True, axis=1)

cols.drop(['esta_online'], inplace=True, axis=1)
# cols.drop(['valor_educacion'], inplace=True, axis=1)
cols.drop(['educacion_int'], inplace=True, axis=1)

cols.drop(['idaviso'], inplace=True, axis=1)
cols.drop(['idtipo_de_trabajo_int'], inplace=True, axis=1)
cols.drop(['idnivel_laboral_int'], inplace=True, axis=1)
cols.drop(['idnombre_area_int'], inplace=True, axis=1)
# cols.drop(['cantidad_postulaciones_x'], inplace=True, axis=1)
# cols.drop(['cantidad_vistas_x'], inplace=True, axis=1)
cols.drop(['month'], inplace=True, axis=1)
cols.drop(['day'], inplace=True, axis=1)
cols.drop(['dayofweek'], inplace=True, axis=1)
cols.drop(['hour'], inplace=True, axis=1)
# cols.drop(['educacion_int'], inplace=True, axis=1)
# cols.drop(['sexo'], inplace=True, axis=1)
# cols.drop(['edad'], inplace=True, axis=1)
# cols.drop(['cantidad_postulaciones_y'], inplace=True, axis=1)
# cols.drop(['cantidad_vistas_y'], inplace=True, axis=1)
cols.drop(['cantidad_interacciones_y'], inplace=True, axis=1)
cols.drop(['cantidad_interacciones_x'], inplace=True, axis=1)
# cols.drop(['post_tipo_de_trabajo'], inplace=True, axis=1)
# cols.drop(['post_nivel_laboral'], inplace=True, axis=1)
# cols.drop(['post_nombre_area'], inplace=True, axis=1)

cols = cols.columns
cols

# X will have the numeric data, Y will have the target class
dfix = df.loc[:,cols]
dfix.head()

min_max_scaler = preprocessing.QuantileTransformer()
x_scaled = min_max_scaler.fit_transform(dfix.values)
dfix = pd.DataFrame(x_scaled)

X = np.array(dfix)
y = np.array(df['se_postulo']) 

pca = PCA(n_components=3)
pca.fit(X)
X = pca.transform(X)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=7506)
X_test

print (cols)
for i in range(1, 2):    
    knn = KNeighborsClassifier(n_neighbors=i)                 # Create KNN classifier
    knn.fit(X_train, y_train)                                 # FIT (this does nothing in normal KNN)
    pred = knn.predict(X_test)                                # predict the response    
    print ("# " + str(i) + ": " + str(accuracy_score(y_test, pred))) # evaluate accuracy

X_full_data = X
pred_full_data = knn.predict(X_full_data)
df['sepostulo'] = pred_full_data

top100_full = pd.read_csv("../test_final_100k.csv", encoding='utf-8')
top100_full = pd.merge(top100_full, df[['idaviso', 'idpostulante', 'sepostulo']], how='left', on=['idaviso','idpostulante'])
top100_full = top100_full[['id', 'sepostulo']]

top100_full.drop_duplicates(inplace=True)

top100_full[top100_full['sepostulo'] == 1].shape

top100_full[top100_full['sepostulo'] == 0].shape

top100_full[top100_full['sepostulo'] == -1].shape

top100_full.to_csv('../data/adri_data/top100k_MinMaxScaler_PCA_KNN_Set19_pp.csv', encoding='utf-8', index=False)

df.loc[df['sepostulo']==1,cols ]

df.loc[df['sepostulo']==0,cols ]

df['knn'] = pred

df.to_csv('../data/adri_data/entrenamiento_postulaciones_vistas17_full_conKNN.csv', encoding='utf-8')

# ['idaviso', 'esta_online', 'valor_educacion', 'sexo', 'edad',
# 'idnombre_area_pref', 'idnivel_laboral_pref', 'idtipo_de_trabajo_pref',
#       'idpostulante_int']
# 1: 0.7996987253024347
# 2: 0.7827130242507732
# 3: 0.8085089702421238
# 4: 0.7990879360523255
# 5: 0.807127794550816
# 6: 0.7996343264358547
# 7: 0.8031782640625525
# 8: 0.796691178235263
# 9: 0.7985391457110355
# 10: 0.7928484458673527
# 11: 0.7939268268878468
# 12: 0.7890681124012218
# 13: 0.7894389058752566
# 14: 0.7850133837644457
# 15: 0.7852569794771612
# 16: 0.7811970509319036
# 17: 0.7811574516288513
# 18: 0.7775815145653436
# 19: 0.7775135157621226

# Con PCA

# 1: 0.7275020095351058
# 2: 0.6959017556143745
# 3: 0.7471506854551911
# 4: 0.7353983004212227
# 5: 0.7551425487243099
# 6: 0.7492110248276931
# 7: 0.7588869740878119
# 8: 0.7556586395925196
# 9: 0.7613225608167314
# 10: 0.7592964262970933
# 11: 0.763162580695787
# 12: 0.7613155186411418
# 13: 0.7639392320608122
# 14: 0.7629553395284357
# 15: 0.764661558071289
# 16: 0.7640579430207513
# 17: 0.7654683901888409
# 18: 0.7647460641783642
# 19: 0.7658476616455955


# Scaler + PCA
# 1: 0.9430038385900837
# 2: 0.9271955596458308
# 3: 0.9211485428528543
# 4: 0.9124645325785192
# 5: 0.9080549225420891
# 6: 0.9025229906958757
# 7: 0.8993202085933071
# 8: 0.895457473902698
# 9: 0.8930752060239184
# 10: 0.8902205086705295
# 11: 0.8884752560032547
# 12: 0.8864817162952354
# 13: 0.8848974276738241
# 14: 0.8832684715296856
# 15: 0.8820350845282566
# 16: 0.8807433480602022
# 17: 0.8798262554088949
# 18: 0.87859407563781
# 19: 0.8776709468347829

# Scaler + PCA sin dia
# 1: 0.9200785343579768
# 2: 0.8960659179960621
# 3: 0.892154894091694

# Index(['idaviso', 'esta_online', 'month', 'day', 'dayofweek', 'hour', 'sexo', 'edad', 'idpostulante_int'], dtype='object')
# 1: 0.9431394507987237
# 2: 0.9274032032649947
# 3: 0.9213219816122722
# 4: 0.912565135107184

# Index(['idaviso', 'month', 'day', 'dayofweek', 'hour', 'sexo', 'edad', 'idpostulante_int'], dtype='object')
# 1: 0.9556886102243557
# 2: 0.9407909853695755
# 3: 0.9342172137365107
# 4: 0.9250760655719233
# 5: 0.9199602257842672
# 6: 0.9137268931082037

# Predigo con todos los datos

top100 = pd.read_csv("../test_final_100k.csv", encoding='utf-8')
top100.head()

toPredict = pd.merge(top100, df, how='inner', on=['idaviso', 'idpostulante'])
toPredict.shape

postulantes = pd.read_csv('../data/adri_data/postulantes2.csv', encoding='utf-8')
# avisos_detalle = pd.read_csv('../data/adri_data/avisos_detalles.csv', encoding='utf-8')

# dfix2 = df.loc[:,cols2]
# dfix2

toPredict2 = pd.merge(top100, postulantes, how='inner', on=['idpostulante'])
# toPredict2.rename({'idaviso_x': 'idaviso'}, inplace=True, axis=1)
# toPredict2.drop('idaviso_y', inplace=True, axis=1)
# toPredict2.drop_duplicates(['idaviso', 'idpostulante'], inplace=True)
# toPredict2.fillna(0, inplace=True, axis=1)
toPredict2['sexo'] = toPredict2['sexo'] * 100
toPredict2.shape

toPredict = pd.concat([toPredict, toPredict2], axis=0)
toPredict.head()

toPredict.drop_duplicates(['idaviso', 'idpostulante'], inplace=True)
toPredict.shape

toPredict = toPredict.loc[:,cols2]
toPredict.head()

toPredict.fillna(-100000, inplace=True)
toPredict.shape

dfix = toPredict.loc[:, cols]
dfix.fillna(-100000, inplace=True)
dfix.head()

min_max_scaler = preprocessing.MinMaxScaler()
x_scaled = min_max_scaler.fit_transform(dfix.values)
dfix = pd.DataFrame(x_scaled)

X = np.array(dfix)
y = np.array(toPredict['se_postulo']) 

pca = PCA(n_components=3)
pca.fit(X)
X = pca.transform(X)

pred = knn.predict(X)       

toPredict['sepostulo'] = pred
toPredict[toPredict['sepostulo'] == 1].shape

toPredict = pd.merge(toPredict, top100, how='inner', on=['idaviso', 'idpostulante'])

# ----------------------------------------------------------

top100_full = toPredict[['id', 'se_postulo', 'sepostulo']]
top100_full.head()

top100_full['se_postulo'] = top100_full['se_postulo'].astype('int')
top100_full['sepostulo'] = top100_full['sepostulo'].astype('int')
top100_full.shape

top100_full['sepostulo'] = top100_full['sepostulo'].apply(lambda x: 0 if x <= 0 else 1)
top100_full['se_postulo'] = top100_full['se_postulo'].apply(lambda x: 0 if x <= 0 else 1)

top100_full[top100_full['sepostulo'] == 1].shape

top100_full[(top100_full['sepostulo'] == 0) & (top100_full['se_postulo'] == 1)]['sepostulo']

top100_full.loc[(top100_full['sepostulo'] == 0) & (top100_full['se_postulo'] == 1), 'sepostulo'] = 1
top100_full[(top100_full['sepostulo'] == 0) & (top100_full['se_postulo'] == 1)]['sepostulo']

top100_full[top100_full['sepostulo'] == 1].shape

top100_full.drop('se_postulo', inplace=True,axis=1)
top100_full.to_csv('../data/adri_data/top100k_Scaler_PCA_KNN_mean.csv', encoding='utf-8', index=False)

df.rename({'id_x' : 'id'}, inplace=True, axis=1)

dfix2 = df

dfix2 = dfix2[['id', 'idaviso', 'idpostulante', 'se_postulo', 'sepostulo']]

# dfix2.drop(['valor_educacion', 'timestamp', 'sexo','edad','idtitulo','idzona','idciudad','idmapacalle',
#             'idtipo_de_trabajo','idnivel_laboral','idnombre_area','iddenominacion_empresa'], inplace=True, axis=1)
dfix2[dfix2['se_postulo'].isnull()].shape

dfix2[dfix2['id'] == 594]

dfix2.fillna(0, inplace=True)

def fillNaWithMean(dfToFill, dfToGet, col):
    mean = dfToGet[dfToGet[col].notnull()][col].mean()
    print('Mean: ' + str(mean))
    # if dtToFill[col]
    dfToFill[col] = mean
    #.fillna(mean, inplace=True)

# toPredict2 = toPredict2.loc[:,cols]

fillNaWithMean(toPredict2, toPredict, 'month')
fillNaWithMean(toPredict2, toPredict, 'day')
fillNaWithMean(toPredict2, toPredict, 'dayofweek')
fillNaWithMean(toPredict2, toPredict, 'hour')

toPredict2.head()
