==> prediction/5-Adri-Arboles.ipynb <==
import pandas as pd
import numpy as np
from sklearn import linear_model
from sklearn import model_selection
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
import matplotlib.pyplot as plt
import seaborn as sb
from sklearn.preprocessing import LabelEncoder
from sklearn.linear_model import LogisticRegression
from sklearn import metrics
from sklearn.datasets import make_blobs
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeClassifier

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.naive_bayes import GaussianNB
from sklearn import metrics
import matplotlib.pyplot as plt
from sklearn.datasets import load_wine
from sklearn.pipeline import make_pipeline

from sklearn.model_selection import train_test_split
#Standard Scaler
from  sklearn.preprocessing  import StandardScaler
#Multi-layer Perceptron
from sklearn.neural_network import MLPClassifier
#PCA
from sklearn.decomposition import PCA
from sklearn import preprocessing

df = pd.read_csv('../data/adri_data/entrenamiento_postulaciones_vistas18_full.csv', encoding='utf-8')
df.head()

df['se_postulo'] = df['se_postulo'].apply(lambda x: 1 if x > 0 else 0)
df['se_postulo'].unique()

df.drop(['titulo', 'descripcion', 'contenido'], inplace=True, axis=1)

df['decil'] = (df['edad'] // 10) * 10

def dropCols(df):
    cols = df.head(0)    
    cols.drop(['idaviso', 'idpostulante', 'se_postulo'], axis=1, inplace=True)    
    
    # cols.drop(['titulo', 'descripcion', 'contenido'
    #           # 'sepostulo'
    #           ], axis=1, inplace=True)    
    
    # cols = cols[['idtipo_de_trabajo_int', 'idnivel_laboral_int', 'idnombre_area_int',
    #    'month', 'day', 'dayofweek', 'hour', 
    #              'sexo', 'edad', 
    #              # 't_0', 'd_0','c_0', 
    #              # 't_1', 'd_1', 'c_1', 
    #              # 't_2', 'd_2', 'c_2', 
    #              # 't_3', 'd_3', 'c_3'
    #             ]]
    return cols.columns

cols = dropCols(df)
cols

dfix = df.loc[:,cols];
dfix.head()

# X will have the numeric data, Y will have the target class

X = np.array(dfix)
y = np.array(df['se_postulo']) 

del dfix

clf = make_pipeline(preprocessing.QuantileTransformer(), # PCA(n_components=3), 
                     DecisionTreeClassifier(max_depth=25))

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=7506)
clf.fit(X_train, y_train)   

clf.score(X_test, y_test)

pred = clf.predict(X_test)

print (cols)
print (accuracy_score(y_test, pred))

# Index(['esta_online', 'idzona_int', 'idciudad_int', 'idmapacalle_int',
#        'idtipo_de_trabajo_int', 'idnivel_laboral_int', 'idnombre_area_int',
#        'iddenominacion_empresa_int', 'month', 'day', 'dayofweek', 'hour',
#        'valor_educacion', 'educacion_int', 'sexo', 'edad', 'idpostulante_int'],
#       dtype='object')
# 0.92958931165349

# Index(['esta_online', 'idzona_int', 'idciudad_int', 'idmapacalle_int',
#        'idtipo_de_trabajo_int', 'idnivel_laboral_int', 'idnombre_area_int',
#        'iddenominacion_empresa_int', 'month', 'day', 'dayofweek', 'hour',
#        'valor_educacion', 'educacion_int', 'sexo', 'edad', 'idpostulante_int'],
#       dtype='object')
# 1.0

# Index(['idtipo_de_trabajo_int', 'idnivel_laboral_int', 'idnombre_area_int',
#        'month', 'day', 'dayofweek', 'sexo', 'edad'],
#       dtype='object')
# 0.8638474066673939

pred_full_data = clf.predict(X)

df['sepostulo'] = pred_full_data

## File a subir 

top100_full = pd.read_csv("../test_final_100k.csv", encoding='utf-8')

top100_full = pd.merge(top100_full, df, how='inner', on=['idaviso','idpostulante'])
top100_full.shape

top100_full = top100_full[['id', 'se_postulo', 'sepostulo']]
top100_full.head()

top100_full[top100_full['sepostulo'] == 1].shape

top100_full[top100_full['sepostulo'] == 0].shape

top100_full[top100_full['sepostulo'] == -1].shape

top100_full['se_postulo'] = top100_full['se_postulo'].astype('int')
top100_full['sepostulo'] = top100_full['sepostulo'].astype('int')
top100_full.head()

top100 = top100_full.groupby('id')['sepostulo'].sum().reset_index()
top100.shape

top100['sepostulo'] = top100['sepostulo'].apply(lambda x: 1 if x > 0 else 0)

top100[top100['sepostulo'] == 0].shape

top100[top100['sepostulo'] == 1].shape

top100['sepostulo'].unique()

top100_full.fillna(0, inplace=True)

top100[top100['sepostulo'] == -1].shape

top100.to_csv('../data/adri_data/top100k_arboles_18_full.csv', encoding='utf-8', index=False)

X_full_data = np.array(dfix)
X_full_data

df

test = top100_full.sort_values('sepostulo', ascending=False)
test.shape

# test.drop_duplicates('id', inplace=True)

test.shape

top100_full.loc[(top100_full['sepostulo'] == 0) & (top100_full['se_postulo'] == 1), 'sepostulo'] = 1
top100_full[(top100_full['sepostulo'] == 0) & (top100_full['se_postulo'] == 1)]['sepostulo']
