==> prediction/6-Adri-Perceptron.ipynb <==
# Realizo la clasificacion usando Neural Network: Multi-layer Perceptron

## Tabla de resultados 

# Aplicando Scaler:                           en Jupyter 0.8105548320964798, en Kaggle 0.83976

# Aplicando Scaler y luego PCA(1 Componente): en Jupyter 0.6473592000460224, en Kaggle 0.86418
# Aplicando Scaler y luego PCA(2 Componente): en Jupyter 0.6878966168264378, en Kaggle 0.86645
# Aplicando Scaler y luego PCA(3 Componente): en Jupyter 0.7500436690357772, en Kaggle 0.84915

# Aplicando PCA(1 Componente) y luego Scaler: en Jupyter 0.7652969755924084, en Kaggle 0.84879
# Aplicando PCA(2 Componente) y luego Scaler: en Jupyter [Probar]          , en Kaggle [Probar]
# Aplicando PCA(3 Componente) y luego Scaler: en Jupyter [Probar]          , en Kaggle [Probar]

## Importo librerias necesarias

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
#Standard Scaler
from  sklearn.preprocessing  import StandardScaler
#Multi-layer Perceptron
from sklearn.neural_network import MLPClassifier
#PCA
from sklearn.decomposition import PCA
from sklearn import preprocessing

import warnings
from math import log
import datetime as dt
import seaborn as sns
warnings.filterwarnings('ignore')


from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.naive_bayes import GaussianNB
from sklearn import metrics
import matplotlib.pyplot as plt
from sklearn.datasets import load_wine
from sklearn.pipeline import make_pipeline


## Leo todos los datos

df = pd.read_csv('../data/adri_data/entrenamiento_postulaciones_vistas19_6_pp.csv', encoding='utf-8')
df.shape

df.drop(['idtipo_de_trabajo_int'], inplace=True, axis=1)
df.drop(['idnivel_laboral_int'], inplace=True, axis=1)
df.drop(['idnombre_area_int'], inplace=True, axis=1)
df.drop(['month'], inplace=True, axis=1)
df.drop(['day'], inplace=True, axis=1)
df.drop(['dayofweek'], inplace=True, axis=1)
df.drop(['hour'], inplace=True, axis=1)
df.drop(['cantidad_postulaciones_y'], inplace=True, axis=1)
df.drop(['cantidad_vistas_y'], inplace=True, axis=1)
df.drop(['cantidad_interacciones_y'], inplace=True, axis=1)
df.drop(['cantidad_postulaciones_x'], inplace=True, axis=1)
df.drop(['cantidad_vistas_x'], inplace=True, axis=1)
df.drop(['cantidad_interacciones_x'], inplace=True, axis=1)

df['se_postulo'] = df['se_postulo'].apply(lambda x: 1 if x >= 1 else 0)

df['sepostulo'] = 0

df = pd.concat([df, df[df['se_postulo'] == 0]], axis=0).sample(frac=1).reset_index(drop=True)

# df['cantidad_interacciones_y'] = df['cantidad_interacciones_y'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_interacciones_x'] = df['cantidad_interacciones_x'].apply(lambda x: 0 if x < 1 else log(x))
df['cantidad_postulaciones_y'] = df['cantidad_postulaciones_y'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_postulaciones_x'] = df['cantidad_postulaciones_x'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_vistas_y'] = df['cantidad_vistas_y'].apply(lambda x: 0 if x < 1 else log(x))
# df['cantidad_vistas_x'] = df['cantidad_vistas_x'].apply(lambda x: 0 if x < 1 else log(x))

# df['cantidad_interacciones_y'] = df['cantidad_interacciones_y'].apply(lambda x: 0 if x <= 0 else 1)
# df['cantidad_interacciones_x'] = df['cantidad_interacciones_x'].apply(lambda x: 0 if x <= 0 else 1)
df['cantidad_postulaciones_y'] = df['cantidad_postulaciones_y'].apply(lambda x: 0 if x <= 0 else 1)
df['cantidad_postulaciones_x'] = df['cantidad_postulaciones_x'].apply(lambda x: 0 if x <= 0 else 1)
df['cantidad_vistas_y'] = df['cantidad_vistas_y'].apply(lambda x: x if x <= 0 else 1)
df['cantidad_vistas_x'] = df['cantidad_vistas_x'].apply(lambda x: x if x <= 0 else 1)

# df['post_tipo_de_trabajo'] = df['post_tipo_de_trabajo'].apply(lambda x: 0 if x <= 0 else log(x))
# df['post_nivel_laboral'] = df['post_nivel_laboral'].apply(lambda x: 0 if x <= 0 else log(x))
# df['post_nombre_area'] = df['post_nombre_area'].apply(lambda x: 0 if x <= 0 else log(x))
df['vist_tipo_de_trabajo'] = df['vist_tipo_de_trabajo'].apply(lambda x: 0 if x <= 0 else log(x))
df['vist_nivel_laboral'] = df['vist_nivel_laboral'].apply(lambda x: 0 if x <= 0 else log(x))
df['vist_nombre_area'] = df['vist_nombre_area'].apply(lambda x: 0 if x <= 0 else log(x))

# df.drop(['idzona', 'idciudad','idmapacalle'], axis=1, inplace=True)
df.head()

df['valor_educacion'] = df['valor_educacion'].astype('int')
df['educacion_int'] = df['educacion_int'].astype('int')
df['esta_online'] = df['esta_online'].apply(lambda x: 1 if x == 1 else 0)
# df['se_postulo'] = df['se_postulo'].apply(lambda x: 1 if x == 1 else -1)

cols = df.head(0)
cols.drop(['se_postulo', 'idpostulante', 'sepostulo'], inplace=True, axis=1)

cols.drop(['esta_online'], inplace=True, axis=1)
# cols.drop(['valor_educacion'], inplace=True, axis=1)
cols.drop(['educacion_int'], inplace=True, axis=1)

cols.drop(['idaviso'], inplace=True, axis=1)
# cols.drop(['idtipo_de_trabajo_int'], inplace=True, axis=1)
# cols.drop(['idnivel_laboral_int'], inplace=True, axis=1)
# cols.drop(['idnombre_area_int'], inplace=True, axis=1)
# cols.drop(['month'], inplace=True, axis=1)
# cols.drop(['day'], inplace=True, axis=1)
# cols.drop(['dayofweek'], inplace=True, axis=1)
# cols.drop(['hour'], inplace=True, axis=1)
# # cols.drop(['educacion_int'], inplace=True, axis=1)
# # cols.drop(['sexo'], inplace=True, axis=1)
# # cols.drop(['edad'], inplace=True, axis=1)
# cols.drop(['cantidad_postulaciones_y'], inplace=True, axis=1)
# cols.drop(['cantidad_vistas_y'], inplace=True, axis=1)
# cols.drop(['cantidad_interacciones_y'], inplace=True, axis=1)
# cols.drop(['cantidad_postulaciones_x'], inplace=True, axis=1)
# cols.drop(['cantidad_vistas_x'], inplace=True, axis=1)
# cols.drop(['cantidad_interacciones_x'], inplace=True, axis=1)
# # cols.drop(['post_tipo_de_trabajo'], inplace=True, axis=1)
# # cols.drop(['post_nivel_laboral'], inplace=True, axis=1)
# # cols.drop(['post_nombre_area'], inplace=True, axis=1)

cols = cols.columns
cols

dfix = df.loc[:,cols]
dfix.head()

X = np.array(dfix)
y = np.array(df['se_postulo']) 

clf = make_pipeline(preprocessing.QuantileTransformer(),  PCA(n_components=1), 
                    MLPClassifier(solver='lbfgs', alpha=1e-5,hidden_layer_sizes=(5, 2), random_state=1)
                   )

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=7506)
clf.fit(X_train, y_train)   

print (cols)
print (clf.score(X_test, y_test))

del y
del X_train
del X_test 
del y_train
del y_test

# Index(['valor_educacion', 'sexo', 'edad', 'vist_tipo_de_trabajo',
#        'vist_nivel_laboral', 'vist_nombre_area', 'post_tipo_de_trabajo',
#        'post_nivel_laboral', 'post_nombre_area'],
#       dtype='object')
# 0.7427826299126615

# Index(['valor_educacion', 'sexo', 'edad', 'vist_tipo_de_trabajo',
#        'vist_nivel_laboral', 'vist_nombre_area', 'post_tipo_de_trabajo',
#        'post_nivel_laboral', 'post_nombre_area'],
#       dtype='object')
# 0.7423145352641793

X_full_data = X
pred_full_data = clf.predict(X_full_data)
df['sepostulo'] = pred_full_data

top100_full = pd.read_csv("../test_final_100k.csv", encoding='utf-8')
top100_full = pd.merge(top100_full, df[['idaviso', 'idpostulante', 'sepostulo']], how='left', on=['idaviso','idpostulante'])
top100_full = top100_full[['id', 'sepostulo']]

top100_full.drop_duplicates(inplace=True)

top100_full[top100_full['sepostulo'] == 1].shape

top100_full[top100_full['sepostulo'] == 0].shape

top100_full[top100_full['sepostulo'] == -1].shape

top100_full.to_csv('../data/adri_data/top100k_QuantileTransformer_PCA1_Perceptron_Set19_pp_2.csv', encoding='utf-8', index=False)

df.loc[df['sepostulo']==0,cols ]

df.loc[df['sepostulo']==1,cols ]
